+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ HEAD
   - fix the issue with incorrectly flattened values in query
   - extend the memory collector to include the allocation order as resource
   - add loading and storing of performance change records
+  - add short printed results for found degradations
 
 0.14.1 (2018-04-19)
 -------------------